refactor(reporting): extract sample-selection model (#239)
